Factored out the common code from studio and publication into dashboard-common.cp.js
